refactor : 패키지 구조 수정, id 타입 수정
